找不到mapper

+++ question-dao/src/main/java/com/paper/question/dao/mapper/SysUserMapper.java
@@ -3,6 +3,7 @@ package com.paper.question.dao.mapper;
 import com.paper.question.domain.entity.SysUser;
 import org.apache.ibatis.annotations.Param;
 
+
 public interface SysUserMapper {
 
     int deleteByPrimaryKey(Long id);
@@ -11,7 +12,7 @@ public interface SysUserMapper {
 
     int insertSelective(SysUser record);
 
-    SysUser selectByPrimaryKey(Long id);
+    SysUser selectByPrimaryKey(Integer id);
 
     int updateByPrimaryKeySelective(SysUser record);
